refactor(team): extract local model launch helpers
