+++ scripts/ci/source-file-size-baseline.json
@@ -40,7 +40,7 @@
     "src/main/ipc/config.ts": 1230,
     "src/main/ipc/configValidation.ts": 1050,
     "src/main/ipc/review.ts": 4498,
-    "src/main/ipc/teams.ts": 5809,
+    "src/main/ipc/teams.ts": 5689,
     "src/main/services/discovery/ProjectScanner.ts": 2111,
     "src/main/services/infrastructure/CliInstallerService.ts": 1690,
     "src/main/services/infrastructure/ConfigManager.ts": 1401,
@@ -63,7 +63,7 @@
     "src/main/services/team/opencode/store/RuntimeStoreManifest.ts": 1210,
     "src/main/services/team/provisioning/TeamProvisioningBootstrapTranscript.ts": 1661,
     "src/main/services/team/provisioning/TeamProvisioningInboxRelayPolicy.ts": 890,
-    "src/main/services/team/provisioning/TeamProvisioningMemberLifecycle.ts": 2679,
+    "src/main/services/team/provisioning/TeamProvisioningMemberLifecycle.ts": 2678,
     "src/main/services/team/provisioning/TeamProvisioningOpenCodeMemberInboxRelay.ts": 1155,
     "src/main/services/team/provisioning/TeamProvisioningOpenCodeRuntimeDelivery.ts": 1009,
     "src/main/services/team/provisioning/TeamProvisioningOpenCodeRuntimeEvidencePolicy.ts": 964,
@@ -77,7 +77,7 @@
     "src/main/services/team/provisioning/TeamProvisioningStreamEvents.ts": 1212,
     "src/main/services/team/ReviewApplierService.ts": 2062,
     "src/main/services/team/ReviewDecisionStore.ts": 1970,
-    "src/main/services/team/runtime/OpenCodeTeamRuntimeAdapter.ts": 1733,
+    "src/main/services/team/runtime/OpenCodeTeamRuntimeAdapter.ts": 1650,
     "src/main/services/team/TaskChangeComputer.ts": 1357,
     "src/main/services/team/TaskChangeLedgerReader.ts": 2159,
     "src/main/services/team/taskLogs/stream/BoardTaskLogStreamService.ts": 2516,
@@ -117,12 +117,12 @@
     "src/renderer/components/team/activity/ActivityItem.tsx": 2032,
     "src/renderer/components/team/activity/ActivityTimeline.tsx": 1073,
     "src/renderer/components/team/activity/LeadThoughtsGroup.tsx": 1186,
-    "src/renderer/components/team/dialogs/CreateTeamDialog.tsx": 3213,
-    "src/renderer/components/team/dialogs/LaunchTeamDialog.tsx": 3416,
-    "src/renderer/components/team/dialogs/providerPrepareDiagnostics.ts": 1653,
-    "src/renderer/components/team/dialogs/ProvisioningProviderStatusList.tsx": 1239,
+    "src/renderer/components/team/dialogs/CreateTeamDialog.tsx": 3182,
+    "src/renderer/components/team/dialogs/LaunchTeamDialog.tsx": 3359,
+    "src/renderer/components/team/dialogs/providerPrepareDiagnostics.ts": 1640,
+    "src/renderer/components/team/dialogs/ProvisioningProviderStatusList.tsx": 1206,
     "src/renderer/components/team/dialogs/TaskDetailDialog.tsx": 1717,
-    "src/renderer/components/team/dialogs/TeamModelSelector.tsx": 3915,
+    "src/renderer/components/team/dialogs/TeamModelSelector.tsx": 3882,
     "src/renderer/components/team/editor/EditorFileTree.tsx": 907,
     "src/renderer/components/team/editor/ProjectEditorOverlay.tsx": 929,
     "src/renderer/components/team/kanban/KanbanBoard.tsx": 919,
@@ -158,6 +158,6 @@
     "src/renderer/utils/teamModelCatalog.ts": 868,
     "src/renderer/utils/teamProvisioningPresentation.ts": 1414,
     "src/shared/types/api.ts": 1158,
-    "src/shared/types/team.ts": 1959
+    "src/shared/types/team.ts": 1958
   }
 }
