+++ src/renderer/components/team/dialogs/CreateTeamDialog.tsx
@@ -17,7 +17,6 @@ import {
   resolveCodexRuntimeSelection,
 } from '@features/codex-runtime-profile/renderer';
 import { useAppTranslation } from '@features/localization/renderer';
-import { useOpenCodeLocalProviders } from '@features/runtime-provider-management/renderer';
 import { api } from '@renderer/api';
 import { ProviderActivityStatusStrip } from '@renderer/components/common/ProviderActivityStatusStrip';
 import {
@@ -105,6 +104,13 @@ import { AnthropicFastModeSelector } from './AnthropicFastModeSelector';
 import { CodexFastModeSelector } from './CodexFastModeSelector';
 import { CodexReconnectPrompt, shouldShowCodexReconnectPrompt } from './CodexReconnectPrompt';
 import {
+  getOrganizationPlacementUnitKindKey,
+  getOrganizationPlacementUnitOptions,
+  getOrganizationUnitLabel,
+} from './createTeamOrganizationPlacement';
+import { ExperimentalLocalModelOverrideCheckbox } from './ExperimentalLocalModelOverride';
+import { resolveExperimentalLocalModelOverride } from './experimentalLocalModelOverrideState';
+import {
   clearInheritedMemberModelsUnavailableForProvider,
   resolveProviderScopedMemberModel,
 } from './memberModelScope';
@@ -152,6 +158,7 @@ import {
 import { TeammateRuntimeCompatibilityNotice } from './TeammateRuntimeCompatibilityNotice';
 import { computeEffectiveTeamModel } from './TeamModelSelector';
 import { getNextSuggestedTeamName } from './teamNameSets';
+import { useOpenCodeLocalModelScope } from './useOpenCodeLocalModelScope';
 import {
   getWorktreeGitBlockingMessage,
   getWorktreeGitControlDisabledReason,
@@ -162,7 +169,6 @@ import {
 import type {
   OrganizationPlacementSelection,
   OrganizationStructurePayload,
-  OrganizationStructureUnitDto,
 } from '@features/organizations/contracts';
 import type { MemberDraft } from '@renderer/components/team/members/MembersEditorSection';
 import type {
@@ -230,91 +236,6 @@ export interface ActiveTeamRef {
   teamName: string;
   displayName: string;
   projectPath: string;
-}
-
-interface OrganizationPlacementUnitOption {
-  unit: OrganizationStructureUnitDto;
-  depth: number;
-}
-
-function compareOrganizationPlacementUnits(
-  left: OrganizationStructureUnitDto,
-  right: OrganizationStructureUnitDto
-): number {
-  if (left.kind === 'organization' && right.kind !== 'organization') return -1;
-  if (right.kind === 'organization' && left.kind !== 'organization') return 1;
-  return getOrganizationUnitLabel(left).localeCompare(getOrganizationUnitLabel(right));
-}
-
-function getOrganizationPlacementUnitOptions(
-  structure: OrganizationStructurePayload | null,
-  organizationId: string
-): OrganizationPlacementUnitOption[] {
-  if (!structure) return [];
-  const units = structure.units.filter(
-    (unit) => unit.organizationId === organizationId && unit.kind !== 'team'
-  );
-  const unitById = new Map(units.map((unit) => [unit.id, unit]));
-  const organizationRootId =
-    structure.organizations.find((organization) => organization.id === organizationId)
-      ?.rootNodeId ?? null;
-  const rootUnit =
-    (organizationRootId ? unitById.get(organizationRootId) : undefined) ??
-    units.find((unit) => unit.kind === 'organization') ??
-    null;
-  const childrenByParentId = new Map<string | null, OrganizationStructureUnitDto[]>();
-
-  for (const unit of units) {
-    const parentId =
-      unit.parentId && unitById.has(unit.parentId)
-        ? unit.parentId
-        : unit.kind !== 'organization' && rootUnit && unit.id !== rootUnit.id
-          ? rootUnit.id
-          : null;
-    const children = childrenByParentId.get(parentId) ?? [];
-    children.push(unit);
-    childrenByParentId.set(parentId, children);
-  }
-
-  for (const children of childrenByParentId.values()) {
-    children.sort(compareOrganizationPlacementUnits);
-  }
-
-  const ordered: OrganizationPlacementUnitOption[] = [];
-  const visited = new Set<string>();
-  const visit = (unit: OrganizationStructureUnitDto, depth: number): void => {
-    if (visited.has(unit.id)) return;
-    visited.add(unit.id);
-    ordered.push({ unit, depth });
-    for (const child of childrenByParentId.get(unit.id) ?? []) {
-      visit(child, depth + 1);
-    }
-  };
-
-  for (const root of childrenByParentId.get(null) ?? []) {
-    visit(root, 0);
-  }
-  for (const unit of units.sort(compareOrganizationPlacementUnits)) {
-    visit(unit, 0);
-  }
-
-  return ordered;
-}
-
-function getOrganizationUnitLabel(unit: OrganizationStructureUnitDto): string {
-  return unit.title ? `${unit.label} - ${unit.title}` : unit.label;
-}
-
-type OrganizationPlacementUnitKindKey =
-  | 'create.organizationPlacement.kind.root'
-  | 'create.organizationPlacement.kind.group';
-
-function getOrganizationPlacementUnitKindKey(
-  unit: OrganizationStructureUnitDto
-): OrganizationPlacementUnitKindKey {
-  return unit.kind === 'organization'
-    ? 'create.organizationPlacement.kind.root'
-    : 'create.organizationPlacement.kind.group';
 }
 
 interface CreateTeamDialogProps {
@@ -606,13 +527,13 @@ export const CreateTeamDialog = ({
   const [prepareWarnings, setPrepareWarnings] = useState<string[]>([]);
   const [prepareChecks, setPrepareChecks] = useState<ProvisioningProviderCheck[]>([]);
   const [allowExperimentalLocalModels, setAllowExperimentalLocalModels] = useState(false);
-  const experimentalLocalModelOverrideAvailable =
-    prepareChecks.some((check) => check.status === 'failed') &&
-    prepareChecks
-      .filter((check) => check.status === 'failed')
-      .every((check) => check.experimentalOverrideAvailable === true);
-  const experimentalLocalModelOverrideEnabled =
-    experimentalLocalModelOverrideAvailable && allowExperimentalLocalModels;
+  const {
+    available: experimentalLocalModelOverrideAvailable,
+    enabled: experimentalLocalModelOverrideEnabled,
+  } = resolveExperimentalLocalModelOverride({
+    checks: prepareChecks,
+    checked: allowExperimentalLocalModels,
+  });
   const providerReadyById = useMemo(
     () => getProvisioningProviderReadyById(prepareChecks),
     [prepareChecks]
@@ -818,21 +739,12 @@ export const CreateTeamDialog = ({
     }
     return statuses;
   }, [effectiveCwd, globalRuntimeProviderStatusById, projectScopedOpenCodeStatus]);
-  const requiresOpenCodeLocalProviderLookup =
-    selectedProviderId === 'opencode' ||
-    members.some((member) => !member.removedAt && member.providerId === 'opencode');
-  const {
-    providers: openCodeLocalProviders,
-    authoritative: openCodeLocalProviderLookupAuthoritative,
-  } = useOpenCodeLocalProviders({
-    enabled: open && requiresOpenCodeLocalProviderLookup,
-    projectPath: effectiveCwd || null,
+  const openCodeLocalModelScope = useOpenCodeLocalModelScope({
+    enabled: open,
+    projectPath: effectiveCwd,
+    selectedProviderId,
+    members,
   });
-  const openCodeLocalProviderIds = useMemo(
-    () =>
-      new Set(openCodeLocalProviders.map((provider) => provider.providerId.trim().toLowerCase())),
-    [openCodeLocalProviders]
-  );
   const memberModelNormalizationDeferredProviderIds = useMemo<ReadonlySet<TeamProviderId>>(
     () => (codexSnapshotPending ? new Set<TeamProviderId>(['codex']) : new Set()),
     [codexSnapshotPending]
@@ -862,14 +774,12 @@ export const CreateTeamDialog = ({
       selectedProviderId,
       runtimeProviderStatusById,
       deferredProviderIds: memberModelNormalizationDeferredProviderIds,
-      openCodeLocalProviderIds,
-      openCodeLocalProviderLookupAuthoritative,
+      ...openCodeLocalModelScope,
     }).members;
   }, [
     memberModelNormalizationDeferredProviderIds,
     members,
-    openCodeLocalProviderIds,
-    openCodeLocalProviderLookupAuthoritative,
+    openCodeLocalModelScope,
     runtimeProviderStatusById,
     selectedProviderId,
     syncModelsWithLead,
@@ -1028,8 +938,7 @@ export const CreateTeamDialog = ({
       selectedProviderId,
       runtimeProviderStatusById,
       deferredProviderIds: memberModelNormalizationDeferredProviderIds,
-      openCodeLocalProviderIds,
-      openCodeLocalProviderLookupAuthoritative,
+      ...openCodeLocalModelScope,
     });
     if (sanitized.changed) {
       setMembers(sanitized.members);
@@ -1037,8 +946,7 @@ export const CreateTeamDialog = ({
   }, [
     memberModelNormalizationDeferredProviderIds,
     members,
-    openCodeLocalProviderIds,
-    openCodeLocalProviderLookupAuthoritative,
+    openCodeLocalModelScope,
     runtimeProviderStatusById,
     selectedProviderId,
     setMembers,
@@ -1179,8 +1087,7 @@ export const CreateTeamDialog = ({
         memberModel: member.model,
         selectedProviderId,
         runtimeProviderStatusById,
-        openCodeLocalProviderIds,
-        openCodeLocalProviderLookupAuthoritative,
+        ...openCodeLocalModelScope,
       });
       if (scopedModel.model) {
         addModel(scopedModel.providerId, scopedModel.model, memberEffort);
@@ -1193,8 +1100,7 @@ export const CreateTeamDialog = ({
   }, [
     effectiveAnthropicRuntimeLimitContext,
     effectiveMemberDrafts,
-    openCodeLocalProviderIds,
-    openCodeLocalProviderLookupAuthoritative,
+    openCodeLocalModelScope,
     runtimeProviderStatusById,
     selectedEffortForCurrentSelection,
     selectedModel,
@@ -3200,26 +3106,13 @@ export const CreateTeamDialog = ({
                   {getProvisioningFailureHint(effectivePrepare.message, prepareChecks, t)}
                 </p>
                 {experimentalLocalModelOverrideAvailable ? (
-                  <div className="mt-2 flex items-start gap-2 pl-6">
-                    <Checkbox
-                      id="create-experimental-local-model"
-                      checked={allowExperimentalLocalModels}
-                      onCheckedChange={(checked) =>
-                        setAllowExperimentalLocalModels(checked === true)
-                      }
-                    />
-                    <div className="space-y-0.5">
-                      <Label
-                        htmlFor="create-experimental-local-model"
-                        className="cursor-pointer text-xs font-medium text-amber-200"
-                      >
-                        {t('launch.prepare.experimentalLocalModelOverride')}
-                      </Label>
-                      <p className="text-[10px] text-[var(--color-text-muted)]">
-                        {t('launch.prepare.experimentalLocalModelOverrideHint')}
-                      </p>
-                    </div>
-                  </div>
+                  <ExperimentalLocalModelOverrideCheckbox
+                    id="create-experimental-local-model"
+                    checked={allowExperimentalLocalModels}
+                    onCheckedChange={setAllowExperimentalLocalModels}
+                    label={t('launch.prepare.experimentalLocalModelOverride')}
+                    hint={t('launch.prepare.experimentalLocalModelOverrideHint')}
+                  />
                 ) : null}
                 {showCodexReconnectPrompt ? (
                   <div className="pl-6">
