+++ src/renderer/components/team/dialogs/TeamModelSelector.tsx
@@ -48,18 +48,13 @@ import {
 } from '@renderer/utils/teamModelAvailability';
 import {
   compareTeamModelVersionsDescending,
-  doesTeamModelCarryProviderBrand,
-  getProviderScopedTeamModelLabel,
   getRuntimeAwareProviderScopedTeamModelLabel,
-  getTeamModelLabel as getCatalogTeamModelLabel,
   getTeamModelSourceBadgeLabel,
-  getTeamProviderLabel as getCatalogTeamProviderLabel,
 } from '@renderer/utils/teamModelCatalog';
 import {
   compareTeamModelRecommendations,
   getTeamModelRecommendation,
 } from '@renderer/utils/teamModelRecommendations';
-import { resolveAnthropicLaunchModel } from '@shared/utils/anthropicLaunchModel';
 import { getAnthropicDefaultTeamModel } from '@shared/utils/anthropicModelDefaults';
 import { parseOpenCodeQualifiedModelRef } from '@shared/utils/opencodeModelRef';
 import {
@@ -94,6 +89,13 @@ import {
   shouldShowOpenCodeNeedsTestBadge,
   shouldShowOpenCodeOverviewStatus,
 } from './teamModelSelectorUi';
+export {
+  computeEffectiveTeamModel,
+  formatTeamModelSummary,
+  getTeamEffortLabel,
+  getTeamModelLabel,
+  getTeamProviderLabel,
+} from './teamModelSummary';
 
 import type { RuntimeLocalProviderListEntryDto } from '@features/runtime-provider-management/contracts';
 import type { CliProviderStatus, TeamProviderId } from '@shared/types';
@@ -907,82 +909,6 @@ function getOpenCodeReadinessMessage(
     return t('modelSelector.openCodeStatus.messages.launchBlocked');
   }
   return t('modelSelector.openCodeStatus.messages.ready');
-}
-
-export function getTeamModelLabel(model: string): string {
-  return getCatalogTeamModelLabel(model) ?? model;
-}
-
-export function getTeamProviderLabel(providerId: TeamProviderId): string {
-  return getCatalogTeamProviderLabel(providerId) ?? 'Anthropic';
-}
-
-export function getTeamEffortLabel(effort: string): string {
-  const trimmed = effort.trim();
-  if (!trimmed) return 'Default';
-  if (trimmed === 'xhigh') return 'XHigh';
-  return trimmed.charAt(0).toUpperCase() + trimmed.slice(1);
-}
-
-export function formatTeamModelSummary(
-  providerId: TeamProviderId,
-  model: string,
-  effort?: string
-): string {
-  const providerLabel = getTeamProviderLabel(providerId);
-  const routeLabel =
-    providerId === 'opencode'
-      ? (getTeamModelSourceBadgeLabel(providerId, model.trim()) ?? providerLabel)
-      : providerLabel;
-  const rawModelLabel = model.trim() ? getTeamModelLabel(model.trim()) : 'Default';
-  const modelLabel = model.trim()
-    ? getProviderScopedTeamModelLabel(providerId, model.trim())
-    : 'Default';
-  const effortLabel = effort?.trim() ? getTeamEffortLabel(effort) : '';
-
-  const modelAlreadyCarriesProviderBrand =
-    doesTeamModelCarryProviderBrand(providerId, rawModelLabel) ||
-    (providerId === 'codex' && model.trim().toLowerCase().startsWith('gpt-'));
-  const providerActsAsBackendOnly =
-    providerId !== 'anthropic' && modelLabel !== 'Default' && !modelAlreadyCarriesProviderBrand;
-
-  const parts = modelAlreadyCarriesProviderBrand
-    ? [modelLabel, effortLabel]
-    : providerActsAsBackendOnly
-      ? [modelLabel, `via ${routeLabel}`, effortLabel]
-      : [providerLabel, modelLabel, effortLabel];
-
-  return parts.filter(Boolean).join(' · ');
-}
-
-/**
- * Computes the effective model string for team provisioning.
- * By default adds [1m] suffix for Opus 1M context.
- * When limitContext=true, returns base model without [1m] (200K context).
- * Standard Sonnet and Haiku selections stay standard context. Explicit Sonnet 1M selections keep
- * their [1m] suffix unless the 200K limit is enabled.
- */
-export function computeEffectiveTeamModel(
-  selectedModel: string,
-  limitContext: boolean,
-  providerId: TeamProviderId = 'anthropic',
-  providerStatus?: Pick<CliProviderStatus, 'providerId' | 'modelCatalog'> | null
-): string | undefined {
-  if (providerId !== 'anthropic') {
-    return selectedModel.trim() || undefined;
-  }
-
-  const catalog =
-    providerStatus?.providerId === 'anthropic' ? (providerStatus.modelCatalog ?? null) : null;
-
-  return (
-    resolveAnthropicLaunchModel({
-      selectedModel,
-      limitContext,
-      availableLaunchModels: catalog?.models.map((model) => model.launchModel),
-      defaultLaunchModel: catalog?.defaultLaunchModel ?? null,
-    }) ?? getAnthropicDefaultTeamModel(limitContext)
-  );
 }
 
 const OpenCodeModelGroupHeader = ({
