+++ src/renderer/components/team/dialogs/teamModelSummary.ts
@@ -1,0 +1,74 @@
+import {
+  doesTeamModelCarryProviderBrand,
+  getProviderScopedTeamModelLabel,
+  getTeamModelLabel as getCatalogTeamModelLabel,
+  getTeamModelSourceBadgeLabel,
+  getTeamProviderLabel as getCatalogTeamProviderLabel,
+} from '@renderer/utils/teamModelCatalog';
+import { resolveAnthropicLaunchModel } from '@shared/utils/anthropicLaunchModel';
+import { getAnthropicDefaultTeamModel } from '@shared/utils/anthropicModelDefaults';
+
+import type { CliProviderStatus, TeamProviderId } from '@shared/types';
+
+export function getTeamModelLabel(model: string): string {
+  return getCatalogTeamModelLabel(model) ?? model;
+}
+
+export function getTeamProviderLabel(providerId: TeamProviderId): string {
+  return getCatalogTeamProviderLabel(providerId) ?? 'Anthropic';
+}
+
+export function getTeamEffortLabel(effort: string): string {
+  const trimmed = effort.trim();
+  if (!trimmed) return 'Default';
+  if (trimmed === 'xhigh') return 'XHigh';
+  return trimmed.charAt(0).toUpperCase() + trimmed.slice(1);
+}
+
+export function formatTeamModelSummary(
+  providerId: TeamProviderId,
+  model: string,
+  effort?: string
+): string {
+  const providerLabel = getTeamProviderLabel(providerId);
+  const routeLabel =
+    providerId === 'opencode'
+      ? (getTeamModelSourceBadgeLabel(providerId, model.trim()) ?? providerLabel)
+      : providerLabel;
+  const rawModelLabel = model.trim() ? getTeamModelLabel(model.trim()) : 'Default';
+  const modelLabel = model.trim()
+    ? getProviderScopedTeamModelLabel(providerId, model.trim())
+    : 'Default';
+  const effortLabel = effort?.trim() ? getTeamEffortLabel(effort) : '';
+  const modelAlreadyCarriesProviderBrand =
+    doesTeamModelCarryProviderBrand(providerId, rawModelLabel) ||
+    (providerId === 'codex' && model.trim().toLowerCase().startsWith('gpt-'));
+  const providerActsAsBackendOnly =
+    providerId !== 'anthropic' && modelLabel !== 'Default' && !modelAlreadyCarriesProviderBrand;
+  const parts = modelAlreadyCarriesProviderBrand
+    ? [modelLabel, effortLabel]
+    : providerActsAsBackendOnly
+      ? [modelLabel, `via ${routeLabel}`, effortLabel]
+      : [providerLabel, modelLabel, effortLabel];
+  return parts.filter(Boolean).join(' · ');
+}
+
+/** Resolves the exact launch model, including Anthropic context-window aliases. */
+export function computeEffectiveTeamModel(
+  selectedModel: string,
+  limitContext: boolean,
+  providerId: TeamProviderId = 'anthropic',
+  providerStatus?: Pick<CliProviderStatus, 'providerId' | 'modelCatalog'> | null
+): string | undefined {
+  if (providerId !== 'anthropic') return selectedModel.trim() || undefined;
+  const catalog =
+    providerStatus?.providerId === 'anthropic' ? (providerStatus.modelCatalog ?? null) : null;
+  return (
+    resolveAnthropicLaunchModel({
+      selectedModel,
+      limitContext,
+      availableLaunchModels: catalog?.models.map((model) => model.launchModel),
+      defaultLaunchModel: catalog?.defaultLaunchModel ?? null,
+    }) ?? getAnthropicDefaultTeamModel(limitContext)
+  );
+}
