+++ src/shared/types/teamProvisioning.ts
@@ -1,0 +1,5 @@
+export type LocalModelLaunchOptions = { allowExperimentalLocalModels?: boolean };
+export type LocalModelIssueMetadata = { experimentalOverrideAvailable?: boolean };
+export type TeamProvisioningModelVerificationMode = 'compatibility' | 'deep';
+export type TeamProvisioningPrepareIssueScope = 'provider' | 'model';
+export type TeamProvisioningPrepareIssueSeverity = 'blocking' | 'warning';
